requirements.txt
PyGithub==1.58.2
PyYAML==6.0
